requirements.txt
requests
beautifulsoup4
pandas
streamlit
plotly
playwright
google-genai
python-dotenv
